# ROX-21558: Cache docker image build