--- a/Dockerfile
+++ b/Dockerfile
@@ -1,13 +1,19 @@
 FROM registry.ci.openshift.org/openshift/release:golang-1.20 AS build
 
-RUN mkdir /rds_ca
+RUN mkdir /src /rds_ca
 ADD https://truststore.pki.rds.amazonaws.com/global/global-bundle.pem /rds_ca/aws-rds-ca-global-bundle.pem
 
-RUN mkdir /src
 WORKDIR /src
+RUN --mount=type=cache,target=/go/pkg/mod/ \
+     --mount=type=bind,source=go.sum,target=go.sum \
+     --mount=type=bind,source=go.mod,target=go.mod \
+      go mod download -x
+
 COPY . ./
 
-RUN make binary
+RUN --mount=type=cache,target=/go/pkg/mod/ \
+    --mount=type=cache,target=/go/.cache/ \
+    make binary
 
 FROM registry.access.redhat.com/ubi8/ubi-minimal:8.9 as standard