--- a/Makefile
+++ b/Makefile
@@ -194,7 +194,6 @@ help:
 	@echo "make openapi/generate            generate openapi modules"
 	@echo "make openapi/validate            validate openapi schema"
 	@echo "make image/build                 build fleet-manager and fleetshard-sync container image"
-	@echo "make image/build/local           build fleet-manager and fleetshard-sync binaries locally, this image supports arm64"
 	@echo "make image/push                  push image"
 	@echo "make setup/git/hooks             setup git hooks"
 	@echo "make secrets/touch               touch all required secret files"
@@ -508,13 +507,6 @@ ifeq ("$(CLUSTER_TYPE)","kind")
 	kind load docker-image $(SHORT_IMAGE_REF)
 endif
 .PHONY: image/build
-
-# Builds the binaries locally and copies them into the image. This build is compatible with arm64.
-image/build/local: GOOS=linux
-image/build/local: DOCKERFILE="Dockerfile.local"
-image/build/local:
-	GOARCH=$(GOARCH) GOOS=$(GOOS) $(MAKE) binary
-	$(MAKE) image/build DOCKERFILE=$(DOCKERFILE)
 
 image/build/probe: GOOS=linux
 image/build/probe: IMAGE_REF="$(external_image_registry)/$(probe_image_repository):$(image_tag)"
@@ -863,15 +855,15 @@ deploy/bootstrap:
 .PHONY: deploy/bootstrap
 
 # Deploy local images fast for development
-deploy/dev-fast: image/build/local deploy/dev-fast/fleet-manager deploy/dev-fast/fleetshard-sync
+deploy/dev-fast: image/build deploy/dev-fast/fleet-manager deploy/dev-fast/fleetshard-sync
 
 deploy/dev-fast/fleet-manager: GOOS=linux
-deploy/dev-fast/fleet-manager: image/build/local
+deploy/dev-fast/fleet-manager: image/build
 	kubectl -n $(ACSCS_NAMESPACE) set image deploy/fleet-manager fleet-manager=$(SHORT_IMAGE_REF) db-migrate=$(SHORT_IMAGE_REF)
 	kubectl -n $(ACSCS_NAMESPACE) delete pod -l application=fleet-manager
 
 deploy/dev-fast/fleetshard-sync: GOOS=linux
-deploy/dev-fast/fleetshard-sync: image/build/local
+deploy/dev-fast/fleetshard-sync: image/build
 	kubectl -n $(ACSCS_NAMESPACE) set image deploy/fleetshard-sync fleetshard-sync=$(SHORT_IMAGE_REF)
 	kubectl -n $(ACSCS_NAMESPACE) delete pod -l application=fleetshard-sync