## Description
This change brings some performance improvements to the `make image/build` by leveraging [docker cache mounts](https://docs.docker.com/build/guide/mounts/) and go [mod] caching.

## Comparison
Here I give a comparison of the execution time between different image build methods, namely:
1. `make image/build` before the change
2. `make image/build` after the change
3. `make image/build/local` before the change 

Tested on my dev laptop.
```
$ docker --version
Docker version 23.0.1, build a5ee5b1dfc
```
Docker runtime: colima
```
$ colima list
PROFILE    STATUS     ARCH      CPUS    MEMORY    DISK     RUNTIME       ADDRESS
default    Running    x86_64    8       16GiB     60GiB    docker+k3s
```

#### Before
```
make image/build  4.73s user 4.25s system 1% cpu 13:58.27 total
```
#### After
Clean docker cache:
```
docker builder prune -af
```

First run
```
make image/build  6.83s user 5.94s system 1% cpu 13:31.22 total
```
Second run (no changes)
```
make image/build  0.45s user 0.63s system 56% cpu 1.912 total
```
Third run (change a comment in `main.go`)
```
make image/build  0.48s user 0.63s system 2% cpu 38.934 total
```
Fourth run (change dependency in `go.mod`)
```
make image/build  0.89s user 1.23s system 1% cpu 2:42.25 total
```

##### Comparison with `make image/build/local

First run (with local go cache):
```
make image/build/local  27.44s user 14.48s system 65% cpu 1:03.80 total
```
Second run (no changes):
```
make image/build/local  8.03s user 8.00s system 72% cpu 22.034 total
```
Third run (change comment in `main.go`)
```
make image/build/local  27.07s user 14.07s system 59% cpu 1:09.00 total
```
Fourth run (change dependency in `go.mod`)
```
make image/build/local  67.26s user 25.45s system 113% cpu 1:22.03 total
```

**With relatively equal runtimes, `image/build/local` becomes redundant, so I removed it to simplify our codebase.**

## Checklist (Definition of Done)
<!-- Please strikethrough options not relevant using two tildes ~~Text~~. Do not delete non relevant options -->
- [ ] Unit and integration tests added
- [ ] Added test description under `Test manual`
- [ ] Documentation added if necessary (i.e. changes to dev setup, test execution, ...)
- [ ] CI and all relevant tests are passing
- [ ] Add the ticket number to the PR title if available, i.e. `ROX-12345: ...`
- [ ] Discussed security and business related topics privately. Will move any security and business related topics that arise to private communication channel.
- [ ] Add secret to app-interface Vault or Secrets Manager if necessary
- [ ] RDS changes were e2e tested [manually](../docs/development/howto-e2e-test-rds.md)
- [ ] Check AWS limits are reasonable for changes provisioning new resources

## Test manual
See "Comparison" section