@kovayur 2024-01-08
@SimonBaeumer added arm64 support